add more ui and confirm component

+++ src/app/components/main-content/main-content.component.ts
@@ -1,55 +1,78 @@
-import { Component, OnInit, ViewChild} from '@angular/core';
-import {MatPaginator, MatPaginatorModule} from '@angular/material/paginator';
-import {MatSort, MatSortModule} from '@angular/material/sort';
-import {MatTableDataSource, MatTableModule} from '@angular/material/table';
-import {MatInputModule} from '@angular/material/input';
-import {MatFormFieldModule} from '@angular/material/form-field';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
+import { MatSort, MatSortModule } from '@angular/material/sort';
+import { MatTableDataSource, MatTableModule } from '@angular/material/table';
+import { MatInputModule } from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
 import { User } from '../../../types';
 import { UserServiceService } from '../../services/user-service.service';
 import { CommonModule } from '@angular/common';
-import {MatIconModule} from '@angular/material/icon';
+import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialog } from '@angular/material/dialog';
 import { Subscription } from 'rxjs';
 import { UserAddEditComponent } from '../user-add-edit/user-add-edit.component';
-import {MatChipsModule} from '@angular/material/chips';
+import { MatChipsModule } from '@angular/material/chips';
+import { ConfirmModalComponent } from '../confirm-modal/confirm-modal.component';
 
 @Component({
   selector: 'app-main-content',
-  imports: [MatTableModule, MatPaginatorModule, MatInputModule, MatFormFieldModule, MatSortModule, CommonModule, MatIconModule, MatButtonModule, MatChipsModule],
+  imports: [
+    MatTableModule,
+    MatPaginatorModule,
+    MatInputModule,
+    MatFormFieldModule,
+    MatSortModule,
+    CommonModule,
+    MatIconModule,
+    MatButtonModule,
+    MatChipsModule,
+  ],
   templateUrl: './main-content.component.html',
-  styleUrl: './main-content.component.scss'
+  styleUrl: './main-content.component.scss',
 })
-export class MainContentComponent implements OnInit{
+export class MainContentComponent implements OnInit {
   users: User[] = [];
   private usersSub!: Subscription;
-  
+
   ngOnInit(): void {
     this.getUsersList();
 
-    this.usersSub = this._userService.getUsersUpdateListener()
-      .subscribe(users => {
+    this.usersSub = this._userService
+      .getUsersUpdateListener()
+      .subscribe((users) => {
         this.dataSource = new MatTableDataSource(users);
         this.dataSource.paginator = this.paginator;
         this.dataSource.sort = this.sort;
       });
   }
   pageSize: number = 10;
-  displayedColumns: string[] = ['firstName', 'lastName', 'createdAt', 'tags', 'email', 'description', 'action'];
+  displayedColumns: string[] = [
+    'firstName',
+    'lastName',
+    'createdAt',
+    'tags',
+    'email',
+    'description',
+    'action',
+  ];
   dataSource!: MatTableDataSource<User>;
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
 
-  constructor(private _userService: UserServiceService, private _dialog: MatDialog) {}
+  constructor(
+    private _userService: UserServiceService,
+    private _dialog: MatDialog
+  ) {}
 
   getUsersList() {
     this._userService.getUsers().subscribe({
       next: (users) => {
-          this._userService.setUsers(users);
+        this._userService.setUsers(users);
       },
-      error: (err) => console.error(err)
-    })
+      error: (err) => console.error(err),
+    });
   }
   ngOnDestroy() {
     if (this.usersSub) {
@@ -58,7 +81,19 @@ export class MainContentComponent implements OnInit{
   }
 
   deleteUser(createdAt: string) {
-    const updatedUsers = this._userService.deleteUser(createdAt);
+    const dialogRef = this._dialog.open(ConfirmModalComponent, {
+      width: '300px',
+      data: {
+        title: 'Deleting user',
+        message: 'Are you sure you want to delete this user?',
+      },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        const updatedUsers = this._userService.deleteUser(createdAt);
+      }
+    });
   }
 
   applyFilter(event: Event) {
@@ -70,12 +105,9 @@ export class MainContentComponent implements OnInit{
     }
   }
 
- 
-
-    openEditForm(data: any) {
-      
-      const dialogRef = this._dialog.open(UserAddEditComponent, {
-        data,
-      });
-    }
+  openEditForm(data: any) {
+    const dialogRef = this._dialog.open(UserAddEditComponent, {
+      data,
+    });
+  }
 }
